Safari-609.1.11.1/LayoutTests/editing/pasteboard/entries-api/datatransfer-items-drop-fileEntry-file-expected.txt

Basic test coverage for fileSystemFileEntry.file()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dataTransfer.items.length is 2
PASS firstEntry.name is "test.txt"
PASS firstEntry.fullPath is "/test.txt"
PASS firstEntry.isFile is true
PASS fileResult.__proto__ is File.prototype
PASS fileResult.name is "test.txt"
PASS fileResult.webkitRelativePath is ""
PASS fileResult.size is 5
PASS fileResult.type is "text/plain"
PASS contentResult is "Hello"
PASS secondEntry.name is "testFiles"
PASS secondEntry.fullPath is "/testFiles"
PASS secondEntry.isDirectory is true
PASS fileEntryResult.name is "file3.txt"
PASS fileEntryResult.fullPath is "/testFiles/subfolder1/file3.txt"
PASS fileEntryResult.isFile is true
PASS file3Result.__proto__ is File.prototype
PASS file3Result.name is "file3.txt"
PASS file3Result.webkitRelativePath is ""
PASS file3Result.size is 2
PASS file3Result.type is "text/plain"
PASS contentResult is "3\n"
PASS successfullyParsed is true

TEST COMPLETE
